+++ four/src/core/mod.rs
@@ -2,7 +2,7 @@ mod account;
 mod arn;
 pub mod convert;
 pub mod function;
-pub mod logical_id;
+mod logical_id;
 pub mod parameter;
 pub mod partition;
 pub mod pseudo;
@@ -17,3 +17,4 @@ pub use arn::{
     arn_builder, AnyArn, Arn, PartialArn, RefNameAccount, RefNamePartition, RefNameRegion,
     RefNameRegionPartition,
 };
+pub use logical_id::{LogicalId, LogicalIdError, LogicalIdentified};
