+++ four/src/iam/resource/access_key.rs
@@ -5,7 +5,7 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
+        LogicalId,
     },
     iam::resource::user::UserName,
 };
