+++ four/src/iam/resource/group.rs
@@ -5,9 +5,8 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
         service::IAM,
-        Arn,
+        Arn, LogicalId,
     },
     iam::{
         resource::{managed_policy::ManagedPolicyArn, policy::Policy},
