+++ four/src/iam/resource/group_policy.rs
@@ -2,7 +2,7 @@ use crate::{
     core::{
         convert::WillBe,
         function::reference::{RefInner, Referenced},
-        logical_id::LogicalId,
+        LogicalId,
     },
     iam::{property::policy_document::PolicyDocument, resource::policy::PolicyName},
 };
