+++ four/src/iam/resource/instance_profile.rs
@@ -5,9 +5,8 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
         service::IAM,
-        Arn,
+        Arn, LogicalId,
     },
     iam::{resource::role::RoleName, util::Path},
 };
