+++ four/src/iam/resource/server_certificate.rs
@@ -8,10 +8,9 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
         service::IAM,
         tag::Tag,
-        Arn,
+        Arn, LogicalId,
     },
     iam::util::Path,
 };
