+++ four/src/iam/resource/user_to_group_addition.rs
@@ -2,7 +2,7 @@ use crate::{
     core::{
         convert::WillBe,
         function::reference::{RefInner, Referenced},
-        logical_id::LogicalId,
+        LogicalId,
     },
     iam::resource::{group::GroupName, user::UserName},
 };
