+++ four/src/lambda/resource/function.rs
@@ -5,9 +5,8 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
         service::Lambda,
-        Arn,
+        Arn, LogicalId,
     },
     iam::resource::role::RoleArn,
     lambda::property::{
