+++ four/src/lambda/resource/layer_version.rs
@@ -9,9 +9,8 @@ use crate::{
             getatt::{Attribute, HaveAtt},
             reference::{RefInner, Referenced},
         },
-        logical_id::LogicalId,
         service::IAM,
-        Arn,
+        Arn, LogicalId,
     },
     lambda::property::{architecture::Architecture, runtime::Runtime},
 };
